Add .model suffix in models' files name
