+++ src/group/data-table.model.ts
@@ -8,13 +8,13 @@ export class DataTable extends Group<Question<any>[][]> {
     code: string,
     description: string,
     questions: Question<any>[][],
-    private _validations: Validation[] = []
+    public validations: Validation[] = []
   ) {
     super(code, description, GroupType.DATATABLE, questions);
   }
 
   public getQuestionByName(name: string): Question<any> {
-    for (const question of this._questions[0]) {
+    for (const question of this.questions[0]) {
       if (question.name === name) {
         return question;
       }
@@ -22,16 +22,12 @@ export class DataTable extends Group<Question<any>[][]> {
   }
 
   public isRequired(): boolean {
-    for (const validation of this._validations) {
+    for (const validation of this.validations) {
       if (ValidationType.REQUIRED === validation.type) {
         return true;
       }
     }
 
     return false;
-  }
-
-  public get validations(): Validation[] {
-    return this._validations;
   }
 }
