+++ src/group/group-type.enum.ts
@@ -1,4 +1,5 @@
 export enum GroupType {
+
   FIELDSET = <any>'group',
   DATATABLE = <any>'datatable'
 }
