+++ src/group/group.model.ts
@@ -1,0 +1,14 @@
+import { GroupType } from '.';
+import { Question } from '../question';
+
+export abstract class Group<QuestionListType> {
+
+  public constructor(
+    public code: string,
+    public description: string,
+    public type: GroupType,
+    public questions: QuestionListType
+  ) { }
+
+  public abstract getQuestionByName(name: string): Question<any>;
+}
