+++ src/question/index.ts
@@ -1,14 +1,14 @@
-export * from './answer';
+export * from './answer.model';
 export * from './question-type.enum';
-export * from './question';
-export * from './checkbox';
-export * from './option';
-export * from './choice';
+export * from './question.model';
+export * from './checkbox.model';
+export * from './option.model';
+export * from './choice.model';
 export * from './dependency-criteria.enum';
-export * from './dependency';
-export * from './radio';
-export * from './select';
-export * from './text';
-export * from './textarea';
+export * from './dependency.model';
+export * from './radio.model';
+export * from './select.model';
+export * from './text.model';
+export * from './textarea.model';
 export * from './dependency.service';
 export * from './select.service';
