+++ src/validation/index.ts
@@ -1,9 +1,9 @@
 export * from './validation-type.enum';
-export * from './validation';
-export * from './email';
-export * from './max';
-export * from './min';
-export * from './max-length';
-export * from './min-length';
-export * from './pattern';
-export * from './required';
+export * from './validation.model';
+export * from './email.model';
+export * from './max.model';
+export * from './min.model';
+export * from './max-length.model';
+export * from './min-length.model';
+export * from './pattern.model';
+export * from './required.model';
