+++ src/validation/max-length.model.ts
@@ -4,12 +4,8 @@ export class MaxLength extends Validation {
 
   public constructor(
     message: string,
-    private _value: number
+    public value: number
   ) {
     super(ValidationType.MAX_LENGTH, message);
-  }
-
-  public get value(): number {
-    return this._value;
   }
 }
