+++ src/validation/max.model.ts
@@ -4,12 +4,8 @@ export class Max extends Validation {
 
   public constructor(
     message: string,
-    private _value: number
+    public value: number
   ) {
     super(ValidationType.MAX, message);
-  }
-
-  public get value(): number {
-    return this._value;
   }
 }
