+++ src/validation/min-length.model.ts
@@ -4,12 +4,8 @@ export class MinLength extends Validation {
 
   public constructor(
     message: string,
-    private _value: number
+    public value: number
   ) {
     super(ValidationType.MIN_LENGTH, message);
-  }
-
-  public get value(): number {
-    return this._value;
   }
 }
