+++ src/validation/min.model.ts
@@ -4,12 +4,8 @@ export class Min extends Validation {
 
   public constructor(
     message: string,
-    private _value: number
+    public value: number
   ) {
     super(ValidationType.MIN, message);
-  }
-
-  public get value(): number {
-    return this._value;
   }
 }
