+++ src/validation/pattern.model.ts
@@ -4,17 +4,13 @@ export class Pattern extends Validation {
 
   public constructor(
     message: string,
-    private _value: string
+    public value: string
   ) {
     super(ValidationType.PATTERN, message);
   }
 
   /** @override */
   public isRequired(): boolean {
-    return ValidationType.PATTERN === this.type && 'true' === this._value.toString();
-  }
-
-  public get value(): string {
-    return this._value;
+    return ValidationType.PATTERN === this.type && 'true' === this.value.toString();
   }
 }
