+++ src/validation/required.model.ts
@@ -4,12 +4,8 @@ export class Required extends Validation {
 
   public constructor(
     message: string,
-    private _requiredTrue: boolean = false
+    public requiredTrue: boolean = false
   ) {
     super(ValidationType.REQUIRED, message);
-  }
-
-  public get requiredTrue(): boolean {
-    return this._requiredTrue;
   }
 }
